testing

+++ src/main/java/com/nelumbo/parqueadero_api/validation/validators/ParqueaderoTieneEspacioValidator.java
@@ -1,7 +1,6 @@
 package com.nelumbo.parqueadero_api.validation.validators;
 
 import com.nelumbo.parqueadero_api.dto.VehicleEntryRequestDTO;
-import com.nelumbo.parqueadero_api.exception.ResourceNotFoundException;
 import com.nelumbo.parqueadero_api.repository.ParkingRepository;
 import com.nelumbo.parqueadero_api.repository.VehicleRepository;
 import com.nelumbo.parqueadero_api.validation.annotations.ParqueaderoTieneEspacio;
@@ -24,7 +23,7 @@ public class ParqueaderoTieneEspacioValidator implements
             return true;
         }
 
-        return parkingRepository.findById(request.parqueaderoId().intValue())
+        return parkingRepository.findById(request.parqueaderoId())
                 .map(parqueadero -> {
                     long vehiculosActivos = vehicleRepository.countActiveVehiclesInParking(Math.toIntExact(request.parqueaderoId()));
                     boolean tieneEspacio = vehiculosActivos < parqueadero.getCapacidad();
